# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deriv Market Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Candle {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
}

export interface SymbolInfo {
  symbol: string;
  name: string;
  multiplier: number;
}

export type FractalLevelType = typeof FractalLevelType[keyof typeof FractalLevelType];


export const FractalLevelType = {
  SUPPORT: 'SUPPORT',
  RESISTANCE: 'RESISTANCE',
} as const;

export interface FractalLevel {
  type: FractalLevelType;
  index: number;
  price: number;
}

export type FractalMarkerType = typeof FractalMarkerType[keyof typeof FractalMarkerType];


export const FractalMarkerType = {
  HH: 'HH',
  HL: 'HL',
  LH: 'LH',
  LL: 'LL',
} as const;

export interface FractalMarker {
  index: number;
  price: number;
  type: FractalMarkerType;
}

export type ChartDataTrend = typeof ChartDataTrend[keyof typeof ChartDataTrend];


export const ChartDataTrend = {
  UPTREND: 'UPTREND',
  DOWNTREND: 'DOWNTREND',
  CONSOLIDATION: 'CONSOLIDATION',
} as const;

export interface ChartData {
  symbol: string;
  trend: ChartDataTrend;
  candles: Candle[];
  markers: FractalMarker[];
  /** @nullable */
  hh_level?: number | null;
  /** @nullable */
  hl_level?: number | null;
  /** @nullable */
  lh_level?: number | null;
  /** @nullable */
  ll_level?: number | null;
  bar_count: number;
  last_updated: string;
}

export type TrendStructureTrend = typeof TrendStructureTrend[keyof typeof TrendStructureTrend];


export const TrendStructureTrend = {
  UPTREND: 'UPTREND',
  DOWNTREND: 'DOWNTREND',
  CONSOLIDATION: 'CONSOLIDATION',
} as const;

export interface TrendStructure {
  trend: TrendStructureTrend;
  pattern?: string;
  last_resistance?: number;
  prev_resistance?: number;
  last_support?: number;
  prev_support?: number;
  bos_level?: number;
  choch_level?: number;
  pullback_level?: number;
  description: string;
}

export interface MACDData {
  macd: number;
  signal: number;
  histogram: number;
  values?: number[];
  signal_values?: number[];
  histogram_values?: number[];
}

export interface TSIData {
  value: number;
  is_oversold: boolean;
  is_overbought: boolean;
  values?: number[];
}

export type SymbolAnalysisState = typeof SymbolAnalysisState[keyof typeof SymbolAnalysisState];


export const SymbolAnalysisState = {
  PULLBACK: 'PULLBACK',
  BOS_CONTINUATION: 'BOS_CONTINUATION',
  CHoCH_REVERSAL: 'CHoCH_REVERSAL',
  IN_TREND: 'IN_TREND',
  CONSOLIDATION: 'CONSOLIDATION',
} as const;

export type SymbolAnalysisTrend = typeof SymbolAnalysisTrend[keyof typeof SymbolAnalysisTrend];


export const SymbolAnalysisTrend = {
  UPTREND: 'UPTREND',
  DOWNTREND: 'DOWNTREND',
  CONSOLIDATION: 'CONSOLIDATION',
} as const;

export interface SymbolAnalysis {
  symbol: string;
  name: string;
  price: number;
  volatility: number;
  state: SymbolAnalysisState;
  trend: SymbolAnalysisTrend;
  fractal_count: number;
  /** @nullable */
  support?: number | null;
  /** @nullable */
  resistance?: number | null;
  /** @nullable */
  bos_level?: number | null;
  /** @nullable */
  choch_level?: number | null;
  description: string;
  structure?: TrendStructure;
  tsi?: TSIData;
  macd?: MACDData;
  last_fractals?: FractalLevel[];
  last_updated?: string;
}

export interface MarketSummary {
  timestamp: string;
  pullback_count: number;
  bos_count: number;
  choch_count: number;
  trending_count: number;
  consolidation_count: number;
  symbols: SymbolAnalysis[];
}