# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deriv Market Analysis API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/chart/{symbol}:
    get:
      operationId: getChartData
      tags: [market]
      summary: Get 1000 candles + fractal markers (HH/HL/LH/LL) for chart display
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chart data with fractal markers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartData"
...
components:
  schemas:
...
    Candle:
      type: object
      properties:
        time:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
      required:
        - time
        - open
        - high
        - low
        - close
...
    FractalMarker:
      type: object
      properties:
        index:
          type: number
        price:
          type: number
        type:
          type: string
          enum: [HH, HL, LH, LL]
      required:
        - index
        - price
        - type
    ChartData:
      type: object
      properties:
        symbol:
          type: string
        trend:
          type: string
          enum: [UPTREND, DOWNTREND, CONSOLIDATION]
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        markers:
          type: array
          items:
            $ref: "#/components/schemas/FractalMarker"
        hh_level:
          type: ["number", "null"]
        hl_level:
          type: ["number", "null"]
        lh_level:
          type: ["number", "null"]
        ll_level:
          type: ["number", "null"]
        bar_count:
          type: number
        last_updated:
          type: string
      required:
        - symbol
        - trend
        - candles
        - markers
        - bar_count
        - last_updated
...