# lib/api-zod/src/generated/types/chartData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deriv Market Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { Candle } from './candle';
import type { ChartDataTrend } from './chartDataTrend';
import type { FractalMarker } from './fractalMarker';

export interface ChartData {
  symbol: string;
  trend: ChartDataTrend;
  candles: Candle[];
  markers: FractalMarker[];
  /** @nullable */
  hh_level?: number | null;
  /** @nullable */
  hl_level?: number | null;
  /** @nullable */
  lh_level?: number | null;
  /** @nullable */
  ll_level?: number | null;
  bar_count: number;
  last_updated: string;
}