# lib/api-zod/src/generated/types/fractalMarker.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deriv Market Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { FractalMarkerType } from './fractalMarkerType';

export interface FractalMarker {
  index: number;
  price: number;
  type: FractalMarkerType;
}